__darwin_suseconds_t fix

+++ rtclient/transcript.c
@@ -100,7 +100,8 @@ void xscript_data_log(ttp_session_t *session, const char *logline)
  *------------------------------------------------------------------------*/
 void xscript_data_start(ttp_session_t *session, const struct timeval *epoch)
 {
-    fprintf(session->transfer.transcript, "START %lu.%06lu\n", epoch->tv_sec, epoch->tv_usec);
+    fprintf(session->transfer.transcript, "START %lu.%06lu\n", 
+             (unsigned long)epoch->tv_sec, (unsigned long)epoch->tv_usec);
     fflush(session->transfer.transcript);
 }
 
@@ -114,7 +115,8 @@ void xscript_data_start(ttp_session_t *session, const struct timeval *epoch)
  *------------------------------------------------------------------------*/
 void xscript_data_stop(ttp_session_t *session, const struct timeval *epoch)
 {
-    fprintf(session->transfer.transcript, "STOP %lu.%06lu\n\n", epoch->tv_sec, epoch->tv_usec);
+    fprintf(session->transfer.transcript, "STOP %lu.%06lu\n\n", 
+             (unsigned long)epoch->tv_sec, (unsigned long)epoch->tv_usec);
     fflush(session->transfer.transcript);
 }
 
@@ -168,6 +170,9 @@ void xscript_open(ttp_session_t *session)
 
 /*========================================================================
  * $Log: transcript.c,v $
+ * Revision 1.6  2008/05/22 17:58:51  jwagnerhki
+ * __darwin_suseconds_t fix
+ *
  * Revision 1.5  2007/12/07 18:10:28  jwagnerhki
  * cleaned away 64-bit compile warnings, used tsunami-client.h
  *
