+++ server/transcript.c
@@ -101,7 +101,8 @@ void xscript_data_log(ttp_session_t *session, const char *logline)
  *------------------------------------------------------------------------*/
 void xscript_data_start(ttp_session_t *session, const struct timeval *epoch)
 {
-    fprintf(session->transfer.transcript, "START %lu.%06lu\n", epoch->tv_sec, epoch->tv_usec);
+    fprintf(session->transfer.transcript, "START %lu.%06lu\n", 
+              (unsigned long)epoch->tv_sec, (unsigned long)epoch->tv_usec);
     fflush(session->transfer.transcript);
 }
 
@@ -115,7 +116,8 @@ void xscript_data_start(ttp_session_t *session, const struct timeval *epoch)
  *------------------------------------------------------------------------*/
 void xscript_data_stop(ttp_session_t *session, const struct timeval *epoch)
 {
-    fprintf(session->transfer.transcript, "STOP %lu.%06lu\n\n", epoch->tv_sec, epoch->tv_usec);
+    fprintf(session->transfer.transcript, "STOP %lu.%06lu\n\n", 
+              (unsigned long)epoch->tv_sec, (unsigned long)epoch->tv_usec);
     fflush(session->transfer.transcript);
 }
 
@@ -164,6 +166,9 @@ void xscript_open(ttp_session_t *session)
 
 /*========================================================================
  * $Log: transcript.c,v $
+ * Revision 1.8  2008/05/22 17:58:51  jwagnerhki
+ * __darwin_suseconds_t fix
+ *
  * Revision 1.7  2008/04/25 10:37:15  jwagnerhki
  * build35 changed 'ipd_current' from int32 to double for much smoother rate changes
  *
